Export the Array newtype; implement manualEval and withDevicePtr.
This commit exposes the Array constructor such that advanced users can
get the C handle and pass it to other code.

manualEval function lets one force the evaluation of the Array. This is
sometimes necessary to avoid memory leaks, especially when running many
operations in a loop.

withDevicePtr lets one access the raw _device_ pointer to help interop
with other libraries. E.g., halide-arrayfire relies on this function to
pass Arrays to Halide.